Add pulp_consumer_id fact
It's useful to know if you're already registered with a pulp server.
If not, (as well as including pulp::consumer and using pulp_register),
I want to be able to declare a yumrepo resource for a repo containing
pulp packages.  On the next puppet run, (host is registered), I can
then declare the same resource with ensure => absent and start using
pulp binds instead.